# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation example",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "articlesAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create yet another article record",
                "operationId": "articleCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Everything is fine"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ArticleStoreRequest": {
                "title": "Example storring request",
                "description": "Some simple request createa as example",
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title of key for storring",
                        "type": "string",
                        "example": "random"
                    },
                    "content_raw": {
                        "title": "Content",
                        "description": "Title of key for storring",
                        "type": "string",
                        "example": "random"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Examples",
            "description": "Some example pages"
        }
    ]
}